Refactorizacion del repositorio de Book, se mueve logica de negocio a los casos de uso.

+++ Sistema_de_gestión_de_una_biblioteca_en_línea/backend/application/use_cases/book/delete_book_use_case.py
@@ -1,6 +1,7 @@
 from domain.ports.book_repository import BookRepository
 from domain.models.book import Book
 from application.ports.book.delete_book import DeleteBook
+from domain.models.exceptions.business_exception import BusinessNotFoundError
 
 class DeleteBookUseCase(DeleteBook):
     
@@ -10,6 +11,9 @@ class DeleteBookUseCase(DeleteBook):
     async def delete_book(self, book_id: str) -> Book:
         
         book = await self.book_repository.find_by_id(book_id)
+        if not book:
+            raise BusinessNotFoundError(book_id, "El ID no existe.")
+        
         await self.book_repository.delete(book)
         return book
